Start using .g2 file as json_serializable output
It's not possible to create a generator which code depends on another generator, so it seems.

If I keep json_serializable as a dependency to build artemis code, the `.dart_tool/.../build.dart` file won't compile because it won't find the generated .g.dart files, as json_serializable wouldn't had been run yet.