Allow scalar mappings to include imports for types
Until now, generated code did not include imports for types mapped to scalars and so it would fail to build.

This allows us to provide a new form of `dart_type` definition: with the type name and a list of imports it might need.

For example:

```yaml
scalar_mappings:
  - graphql_type: BigDecimal
    dart_type:
      name: Decimal
      imports:
        - package:decimal/decimal.dart
```

The old form, a string with the type name, is still supported for backward compatibility.

Closes #5.