Add .graphql extension to generated output files
So, for some time I've been having an issue with artemis when used in conjunction with other generators that depends of json_serializable (or actually combining_builder, I'm not sure). Basically, the file dependency tree on `build_runner` package was not being updated with the new files artemis was generating.

Then I realized artemis output files weren't being considered because it wouldn't match  `build_extensions`. Sadly, I can't use a build extension used by  another generator (like `.g.dart`) neither, of course, only `.dart` files. A custom unique extension must be used (and that's why it was configured as `.graphql.dart`).

So this PR adds `.graphql.` to generated files, in a non-breaking change way: a "forwarder" file will be generated to make it retro-compatible when a configuration output doesn't include `.graphql.dart` on  the path.